+++ include/notcurses/direct.h
@@ -207,12 +207,8 @@ API int ncdirect_stop(struct ncdirect* nc);
 API ncdirectv* ncdirect_render_frame(struct ncdirect* n, const char* filename,
                                      ncblitter_e blitter, ncscale_e scale);
 
-// Takes the result of ncdirect_render_frame() and writes it to the output. The
-// 'align', 'blitter', and 'scale' arguments must be the same as those passed
-// to ncdirect_render_frame().
-API int ncdirect_raster_frame(struct ncdirect* n, ncdirectv* faken,
-                              ncalign_e align, ncblitter_e blitter,
-                              ncscale_e scale);
+// Takes the result of ncdirect_render_frame() and writes it to the output.
+API int ncdirect_raster_frame(struct ncdirect* n, ncdirectv* ncdv, ncalign_e align);
 
 #undef API