+++ NEWS.md
@@ -4,6 +4,7 @@ rearrangements of Notcurses.
 * 2.1.3 (not yet released)
   * `ncdirect_styles_{set, on, off}()` have been deprecated in favor of
     `ncdirect_{set, on, off}_styles()`, to match `ncplane_` equivalents.
+  * `ncdirect_raster_frame()` no longer requires `blitter` nor `scale`.
 
 * 2.1.2 (2020-12-25)
   * Add `notcurses_linesigs_enable()` and `notcurses_linesigs_disable()`.